+++ packages/apollo-link-http/src/__tests__/httpLink.ts
@@ -231,6 +231,41 @@ describe('HttpLink', () => {
       done();
     });
   });
+  it('adds headers to the request from the setup', done => {
+    const variables = { params: 'stub' };
+    const link = createHttpLink({
+      uri: 'data',
+      headers: { authorization: '1234' },
+    });
+
+    execute(link, { query: sampleQuery, variables }).subscribe(result => {
+      const headers = fetchMock.lastCall()[1].headers;
+      expect(headers.authorization).toBe('1234');
+      expect(headers['content-type']).toBe('application/json');
+      expect(headers.accept).toBe('*/*');
+      done();
+    });
+  });
+  it('prioritizes context headers over setup headers', done => {
+    const variables = { params: 'stub' };
+    const middleware = new ApolloLink((operation, forward) => {
+      operation.setContext({
+        headers: { authorization: '1234' },
+      });
+      return forward(operation);
+    });
+    const link = middleware.concat(
+      createHttpLink({ uri: 'data', headers: { authorization: 'no user' } }),
+    );
+
+    execute(link, { query: sampleQuery, variables }).subscribe(result => {
+      const headers = fetchMock.lastCall()[1].headers;
+      expect(headers.authorization).toBe('1234');
+      expect(headers['content-type']).toBe('application/json');
+      expect(headers.accept).toBe('*/*');
+      done();
+    });
+  });
   it('adds headers to the request from the context on an operation', done => {
     const variables = { params: 'stub' };
     const link = createHttpLink({ uri: 'data' });
@@ -266,6 +301,47 @@ describe('HttpLink', () => {
       done();
     });
   });
+  it('adds creds to the request from the setup', done => {
+    const variables = { params: 'stub' };
+    const link = createHttpLink({ uri: 'data', credentials: 'same-team-yo' });
+
+    execute(link, { query: sampleQuery, variables }).subscribe(result => {
+      const creds = fetchMock.lastCall()[1].credentials;
+      expect(creds).toBe('same-team-yo');
+      done();
+    });
+  });
+  it('prioritizes creds from the context over the setup', done => {
+    const variables = { params: 'stub' };
+    const middleware = new ApolloLink((operation, forward) => {
+      operation.setContext({
+        credentials: 'same-team-yo',
+      });
+      return forward(operation);
+    });
+    const link = middleware.concat(
+      createHttpLink({ uri: 'data', credentials: 'error' }),
+    );
+
+    execute(link, { query: sampleQuery, variables }).subscribe(result => {
+      const creds = fetchMock.lastCall()[1].credentials;
+      expect(creds).toBe('same-team-yo');
+      done();
+    });
+  });
+  it('adds fetcherOptions to the request from the setup', done => {
+    const variables = { params: 'stub' };
+    const link = createHttpLink({
+      uri: 'data',
+      fetcherOptions: { signal: 'foo' },
+    });
+
+    execute(link, { query: sampleQuery, variables }).subscribe(result => {
+      const signal = fetchMock.lastCall()[1].signal;
+      expect(signal).toBe('foo');
+      done();
+    });
+  });
   it('adds fetcherOptions to the request from the context', done => {
     const variables = { params: 'stub' };
     const middleware = new ApolloLink((operation, forward) => {
@@ -277,6 +353,26 @@ describe('HttpLink', () => {
       return forward(operation);
     });
     const link = middleware.concat(createHttpLink({ uri: 'data' }));
+
+    execute(link, { query: sampleQuery, variables }).subscribe(result => {
+      const signal = fetchMock.lastCall()[1].signal;
+      expect(signal).toBe('foo');
+      done();
+    });
+  });
+  it('prioritizes context over setup', done => {
+    const variables = { params: 'stub' };
+    const middleware = new ApolloLink((operation, forward) => {
+      operation.setContext({
+        fetcherOptions: {
+          signal: 'foo',
+        },
+      });
+      return forward(operation);
+    });
+    const link = middleware.concat(
+      createHttpLink({ uri: 'data', fetcherOptions: { signal: 'bar' } }),
+    );
 
     execute(link, { query: sampleQuery, variables }).subscribe(result => {
       const signal = fetchMock.lastCall()[1].signal;