+++ packages/apollo-link-http/src/httpLink.ts
@@ -83,9 +83,17 @@ export interface FetchOptions {
   uri?: string;
   fetch?: GlobalFetch['fetch'];
   includeExtensions?: boolean;
+  credentials?: string;
+  headers?: any;
+  fetcherOptions?: any;
 }
 export const createHttpLink = (
-  { uri, fetch: fetcher, includeExtensions }: FetchOptions = {},
+  {
+    uri,
+    fetch: fetcher,
+    includeExtensions,
+    ...requestOptions,
+  }: FetchOptions = {},
 ) => {
   // dev warnings to ensure fetch is present
   warnIfNoFetch(fetcher);
@@ -119,9 +127,12 @@ export const createHttpLink = (
           throw parseError;
         }
 
+        let options = fetcherOptions;
+        if (requestOptions.fetcherOptions)
+          options = { ...requestOptions.fetcherOptions, ...options };
         const fetchOptions = {
           method: 'POST',
-          ...fetcherOptions,
+          ...options,
           headers: {
             // headers are case insensitive (https://stackoverflow.com/a/5259004)
             accept: '*/*',
@@ -130,7 +141,15 @@ export const createHttpLink = (
           body: serializedBody,
         };
 
+        if (requestOptions.credentials)
+          fetchOptions.credentials = requestOptions.credentials;
         if (credentials) fetchOptions.credentials = credentials;
+
+        if (requestOptions.headers)
+          fetchOptions.headers = {
+            ...fetchOptions.headers,
+            ...requestOptions.headers,
+          };
         if (headers)
           fetchOptions.headers = { ...fetchOptions.headers, ...headers };