+++ packages/apollo-link-http/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change log
 
 ### vNEXT
+- support setting credentials, headers, and fetcherOptions in the setup of the link
 - removed sending of context to the server and allowed opt-in of sending extensions